REFACTORING: DELETING ALL LIGHTS
Only leaving one Directional Lights. Ret to be covered with decals, Post Processing and particles.

Due to having too many lights, they were not properly shown in Unity, so we increased the intensity too much to see them, and when baked, the whole scene was over burned.

As it is now difficult to find the right intensity amount, it has been preferred to add the necessary light zones with decals.